EVALUATION/requirements.txt
# Core ML frameworks
torch==2.5.1
transformers==4.55.0

# LangChain (minimal requirements)
langchain==0.3.27
langchain-community==0.3.30
langchain-huggingface

# Data and evaluation
datasets==4.0.0
sacrebleu==2.5.1
pandas==2.3.1

# Utilities
pyyaml==6.0.2
tqdm==4.67.1

# Optional: Only install if using API providers
# langchain-openai>=0.0.5
# langchain-anthropic>=0.1.0
# openai>=1.0.0
# anthropic>=0.18.0